Reorganize auth implementations
