+++ server/routes/oauth.js
@@ -1,19 +1,93 @@
 const passport = require('passport');
+const PassportGoogleStrategy = require('passport-google-oauth20').Strategy;
 const router = require('express').Router();
 const config = require('../lib/config');
+const logger = require('../lib/logger');
+const checkWhitelist = require('../lib/check-whitelist.js');
+const usersUtil = require('../models/users.js');
 const baseUrl = config.get('baseUrl');
 
-router.get(
-  '/auth/google',
-  passport.authenticate('google', { scope: ['profile email'] })
-);
+const googleClientId = config.get('googleClientId');
+const googleClientSecret = config.get('googleClientSecret');
+const publicUrl = config.get('publicUrl');
 
-router.get(
-  '/auth/google/callback',
-  passport.authenticate('google', {
-    successRedirect: baseUrl + '/',
-    failureRedirect: baseUrl + '/signin'
-  })
-);
+async function passportGoogleStrategyHandler(
+  accessToken,
+  refreshToken,
+  profile,
+  done
+) {
+  const email = profile && profile._json && profile._json.email;
+
+  if (!email) {
+    return done(null, false, {
+      message: 'email not provided from Google'
+    });
+  }
+
+  try {
+    let [openAdminRegistration, user] = await Promise.all([
+      usersUtil.adminRegistrationOpen(),
+      usersUtil.findOneByEmail(email)
+    ]);
+
+    if (user) {
+      user.signupDate = new Date();
+      const newUser = await usersUtil.save(user);
+      newUser.id = newUser._id;
+      return done(null, newUser);
+    }
+    const whitelistedDomains = config.get('whitelistedDomains');
+    if (openAdminRegistration || checkWhitelist(whitelistedDomains, email)) {
+      const newUser = await usersUtil.save({
+        email,
+        role: openAdminRegistration ? 'admin' : 'editor',
+        signupDate: new Date()
+      });
+      newUser.id = newUser._id;
+      return done(null, newUser);
+    }
+    // at this point we don't have an error, but authentication is invalid
+    // per passport docs, we call done() here without an error
+    // instead passing false for user and a message why
+    return done(null, false, {
+      message: "You haven't been invited by an admin yet."
+    });
+  } catch (error) {
+    done(error, null);
+  }
+}
+
+if (config.googleAuthConfigured()) {
+  logger.info('Enabling Google authentication strategy.');
+
+  // Register the passport strategy
+  passport.use(
+    new PassportGoogleStrategy(
+      {
+        clientID: googleClientId,
+        clientSecret: googleClientSecret,
+        callbackURL: publicUrl + baseUrl + '/auth/google/callback',
+        // This option tells the strategy to use the userinfo endpoint instead
+        userProfileURL: 'https://www.googleapis.com/oauth2/v3/userinfo?alt=json'
+      },
+      passportGoogleStrategyHandler
+    )
+  );
+
+  // Add routes for oauth flow
+  router.get(
+    '/auth/google',
+    passport.authenticate('google', { scope: ['profile email'] })
+  );
+
+  router.get(
+    '/auth/google/callback',
+    passport.authenticate('google', {
+      successRedirect: baseUrl + '/',
+      failureRedirect: baseUrl + '/signin'
+    })
+  );
+}
 
 module.exports = router;
